test(mbt): Refactor MBT tests after spec cleanup (#146)
* test(mbt): Revise Emerald's Quint specification

Originally, the spec has been extracted from app.rs with AI assistance.
It was a good starting point but, it ended up carrying a lot of concrete
aspects of the system into the model.

In this patch, I've done a pass simplifying the model, raising its level
of abstraction, and cleaning up the spec to make it easier to understand
and work with.

* test(mbt): Refactor MBT tests after spec cleanup

I've broken down the driver into multiple files to aid with maintenance.
It's now fully implemented to run a single height commit (happy path).